Change TotalHoursGoal type to double and display chart for goals

+++ Models/CodingGoal.cs
@@ -5,5 +5,5 @@ public class CodingGoal
     public int Id { get; init; }
     public DateTime StartTime { get; init; }
     public DateTime EndTime { get; init; }
-    public int TotalHoursGoal { get; init; }
+    public double TotalHoursGoal { get; init; }
 }
